refactor: add corresponding folders for each component
