+++ app/[locale]/(main)/rest/[...slug]/page.tsx
@@ -1,4 +1,4 @@
-import { RestResponce } from '@/components/rest-form/responce';
+import { RestResponce } from '@/components/rest-form/responce/responce';
 
 export default function RestClientPage() {
   return <RestResponce />;
