+++ app/[locale]/(main)/rest/layout.tsx
@@ -1,8 +1,8 @@
-import { Body } from '@/components/rest-form/body';
-import { GeneratedCode } from '@/components/rest-form/generated-code';
-import { RestHeaders } from '@/components/rest-form/headers';
-import { RestMain } from '@/components/rest-form/main';
-import { SendButton } from '@/components/rest-form/send-button';
+import { Body } from '@/components/rest-form/body/body';
+import { GeneratedCode } from '@/components/rest-form/generated-code/generated-code';
+import { RestHeaders } from '@/components/rest-form/headers/headers';
+import { RestMain } from '@/components/rest-form/main/main';
+import { SendButton } from '@/components/rest-form/send-button/send-button';
 import { Locale } from '@/type';
 
 import { generateStaticParams } from '../../layout';
