+++ components/index.tsx
@@ -1,0 +1,4 @@
+export { Footer } from '@/components/footer/footer';
+export { Header } from '@/components/header/header';
+export { LanguageSelect } from '@/components/language-select/language-select';
+export { Wrapper } from '@/components/wrapper/wrapper';
