+++ components/rest-form/header-item/header-item.tsx
@@ -1,6 +1,6 @@
 import { Header } from '@/type';
 
-import { Button } from '../ui/button';
+import { Button } from '../../ui/button';
 
 export const HeaderItem = ({
   header,
