+++ components/rest-form/responce/responce.tsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 
 import { RootState } from '@/store/store';
 
-import { BodyEditor } from './json-editor';
+import { BodyEditor } from '../json-editor/json-editor';
 
 export const RestResponce = () => {
   const { error, loading, response } = useSelector((state: RootState) => state.restRequest);
